Allow for update of repositories with submodules
Although not encouraged, it is possible to import a subrepo that
contains submodules. Despite the initial import working flawlessly, any
request to update the submodule in the subrepo through a repeated import
will fail. The reason for this failure is that the submodule is
represented internally as a commit object and so far we omitted these
objects from the 'diff' command removing the previous state before
applying the new one.
This change fixes the problem by adding commit objects to the list of
objects to recognize.