Allow for importing of remote repository commits by tag
So far we have prohibited importing of remote repository commits
specified by tag. That was less intentional than it was incidental: By
virtue of our commit resolution function, which, at the time, tried to
play very safe, we raised an exception when specifying a tag as the
source commit.
The rationale behind this behavior was that we only supported importing
by branch or SHA1, and wanted to make sure that a branch does not
resolve to a local branch but indeed a remote one. However, since then
we have added additional checking to make sure that a commit does indeed
belong to the given remote repository.
As such, this change removes the aforementioned check in the remote
commit resolution function -- allowing for imports by tag.