Add support for only reimporting imports of a specific remote repository
When a repository gets bigger and has a greater number of subrepository
imports, it is sometimes desired to not perform a reimport on all of
those import commits but just ones for a particular remote repository.
So far this was not possible from within the program and would have to
be scripted on top.
With this change we add native support for such workflows. It adds a new
parameter, -r/--remote, to the reimport command which, if passed in
along with the name of a remote repository, will cause only imports of
this remote repository to be reimported.